requirements.txt
biopython>=1.83
pybliometrics>=3.6
requests>=2.31
pandas>=2.1
rispy>=0.10
rapidfuzz>=3.6
unidecode>=1.3
pyyaml>=6.0
tqdm>=4.66
